+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ This project adheres to `Semantic Versioning http://semver.org/>`_.
 UNRELEASED CHANGES
 ******************
 * CIP-2447 Added FpgaDieVoltage[0-6] Attributes in TalonBoard Device to read from the FPGA Die Voltage Sensors
+* MAP-115 Updated MCS overview Taranta dashboard to include more info LRUs, sim mode and updates to the DISH ID
+* MAP-116 Change initial board IP loading so it is set to an explicitly placeholder value until a HW config file is applied
 
 0.15.1
 ******